Revert "Fix for category sync conflict"
