+++ Model/Sync/Catalog/Processor/Main.php
@@ -64,11 +64,6 @@ class Main extends AbstractJobs
     private $syncByOrderFlag = false;
 
     /**
-     * @var boolean
-     */
-    protected $normalSync = true;
-
-    /**
      * AbstractJobs constructor.
      * @param AppEmulation $appEmulation
      * @param ResourceConnection $resourceConnection
@@ -611,11 +606,18 @@ class Main extends AbstractJobs
     }
 
     /**
-     * @param bool $flag
      * @return void
      */
-    public function setNormalSyncFlag($flag)
+    public function setSyncByOrderFlag()
     {
-        $this->normalSync = $flag;
+        $this->syncByOrderFlag = true;
+    }
+
+    /**
+     * @return bool
+     */
+    public function getSyncByOrderFlag()
+    {
+        return $this->syncByOrderFlag;
     }
 }
