+++ Model/Sync/Orders/Processor.php
@@ -165,10 +165,6 @@ class Processor extends Main
         $storeId = $order->getStoreId();
         $this->emulateFrontendArea((int)$storeId);
         $this->currentStoreId = $this->config->getStoreId();
-        if (!$this->config->isOrdersSyncActive()) {
-            $this->stopEnvironmentEmulation();
-            return;
-        }
         $this->yotpoOrdersLogger->info(
             __(
                 'Process order for Magento Store ID: %1, Name: %2',
@@ -422,14 +418,12 @@ class Processor extends Main
      * @param Order $order
      * @param bool $isYotpoSyncedOrder
      * @param array<mixed> $yotpoSyncedOrders
-     * @param bool $realTImeSync
      * @return array<mixed>|DataObject
      * @throws LocalizedException
      * @throws NoSuchEntityException
      */
-    public function syncOrder($order, $isYotpoSyncedOrder, $yotpoSyncedOrders, $realTImeSync = false)
+    public function syncOrder($order, $isYotpoSyncedOrder, $yotpoSyncedOrders)
     {
-        $orderIds = [];
         $incrementId = $order->getIncrementId();
         $orderId = $order->getEntityId();
         $dataType = $isYotpoSyncedOrder ? 'update' : 'create';
@@ -474,10 +468,7 @@ class Processor extends Main
             if ($yotpoOrderId) {
                 $response->setData('yotpo_id', $yotpoOrderId);
             }
-            if ($realTImeSync) {
-                $orderIds[] = $orderId;
-                $this->updateOrderAttribute($orderIds, self::SYNCED_TO_YOTPO_ORDER, 1);
-            }
+
             $this->yotpoOrdersLogger->info('Orders sync - success - ' . $orderId, []);
         } elseif ($response->getData('status') == 409) {//order already exists in Yotpo and not in custom table
             $response = $this->yotpoCoreSync->sync(
@@ -491,7 +482,7 @@ class Processor extends Main
             if (array_key_exists($orderId, $yotpoSyncedOrders)) {
                 unset($yotpoSyncedOrders[$orderId]);
             }
-            $response = $this->syncOrder($order, false, $yotpoSyncedOrders, $realTImeSync);
+            $response = $this->syncOrder($order, false, $yotpoSyncedOrders);
         }
         if ($this->isCommandLineSync && !$immediateRetry) {
             // phpcs:ignore
@@ -511,7 +502,7 @@ class Processor extends Main
     {
         $unSyncedProductIds = $this->data->getUnSyncedProductIds($productIds, $order);
         if ($unSyncedProductIds) {
-            $this->catalogProcessor->setNormalSyncFlag(false);
+            $this->catalogProcessor->setSyncByOrderFlag();
             $sync = $this->catalogProcessor->process($unSyncedProductIds, $order);
             $this->emulateFrontendArea($this->currentStoreId);
             return $sync;
